Adds the SQL Alchemy entity -> domain mapping
It inverts the logic in the basic SQL Alchemy basic uses where we
define classes that maps to the database schema, it is bad because
it makes our code coupled with the ORM declaration method.

Using the MetaData and Table declaration we can declare the database
schema and finally maps it to our domain classes which makes a clear
boundary between system layers and the infrastructure.

---

Also, but not least this commit adds the basic infrastructure to
connect to an in memory database providing a global session over the
test class.