Adds the UnitOfWork concept to the project
This class will wrap all codes that interacts with the database
infrastructure. It is responsible to handles possible errors and
handle the commit session when the block is finished.

For exemple:

```
with UnitOfWork(session_factory) as u:
  u.account.some_job()
  ..
  ..
  ..
```

The context block will commit the transaction and close the connection
with the session. So all transactions are independents at application
level.