Add TransactionTypeEnum to the Transaction domain
This will help to us distingue what kind of transaction were made by
the users.

When we look at ours bank account we should know what type of transaction
occours, if it is a deposit or a  withdraw for example.

This model can be upgraded to handle transfers for example.