Adds the FastAPI as requirement
It is a good oportunit tu test the FastAPI framework. Choosing the
Hexagon architecture allow to do this thing. If in the future we need
to move out the FastAPI all our logic model are implemented at domain
and services layers. Peharps it will never be a problem.

FastAPI looks a good new alternative to Flask or Bottle, Django never
was a option it is too opinated (if you don't know well of course).

Anyway choosing a micro framework and a new framework is on my on and
it looks a good choose for now.